+++ src/main/java/swp490/spa/repositories/RatingRepository.java
@@ -18,5 +18,5 @@ public interface RatingRepository extends JpaRepository<Rating, Integer> {
     List<Rating> findByStatusRatingAndExpireTime(StatusRating statusRating, Date date);
 
     @Query("FROM Rating r WHERE r.rate = ?1 AND r.bookingDetailStep.bookingDetail.booking.spa.id = ?2")
-    List<Rating> findByRateAndSpa(Double rate, Integer spaId);
+    List<Rating> findByRateAndSpa(Integer rate, Integer spaId);
 }
